+++ packages/rpc_dart_transports/lib/src/transports/websocket/websocket_responder_transport.dart
@@ -26,7 +26,13 @@ class RpcWebSocketResponderTransport extends RpcWebSocketTransportBase {
   ///
   /// [channel] WebSocket канал для коммуникации
   /// [logger] Опциональный логгер для отладки
-  RpcWebSocketResponderTransport(super.channel, {super.logger});
+  RpcWebSocketResponderTransport(
+    super.channel, {
+    super.logger,
+    super.chunkSizeBytes,
+    super.maxChunkedMessageBytes,
+    super.enableChunking,
+  });
 
   @override
   bool get supportsZeroCopy => false;